requirements.txt
boto3
numpy
Pillow
requests
awslambdaric
opencv-python-headless
facenet-pytorch
torch
torchvision
